+++ rampt/gui/helpers.py
@@ -29,13 +29,13 @@ def evaluate_dialog(state, payload_pressed, option_list):
 
 
 ## Working directory
-def change_work_dir_root(gui, new_root: StrPath = None):
+def change_work_dir_root(gui, new_root: StrPath = None, logger: Logger = Logger()):
     global work_dir_root
     if new_root:
         if os.path.isdir(new_root):
             work_dir_root = os.path.normpath(new_root)
         else:
-            error(message=f"{new_root} is not a valid directory", error_type=ValueError)
+            logger.error(message=f"{new_root} is not a valid directory", error_type=ValueError)
     else:
         work_dir_root = gui._get_config("upload_folder", tempfile.gettempdir())
 
@@ -61,10 +61,10 @@ def download_directory(state, dir):
             download(state, content=os.path.join(root, file), name=os.path.basename(file))
 
 
-def download_data_node_files(state, data_node_name, files_attribute: str = "scenario.data_nodes"):
+def download_data_node_files(state, data_node_name, files_attribute: str = "scenario.data_nodes", logger: Logger = Logger()):
     entries = get_attribute_recursive(state, files_attribute)
     if data_node_name:
-        log("Ready for reading: " + str(entries[data_node_name].is_ready_for_reading))
+        logger.log("Ready for reading: " + str(entries[data_node_name].is_ready_for_reading))
         entries = entries[data_node_name].read()
 
     entries = entries if isinstance(entries, list) else [entries]