+++ rampt/gui/pages/analysis/analysis.py
@@ -9,6 +9,6 @@ analysis_params = Analysis_Runner()
 
 
 # TODO: Implement analysis
-def create_analysis():
+def create_analysis(logger: Logger = Logger()):
     tgb.text("###### File selection", mode="markdown")
-    create_file_selection(process="analysis", out_node="analysis_data_paths")
+    create_file_selection(process="analysis", out_node="analysis_data_paths", logger=logger)