+++ rampt/gui/pages/analysis/summary.py
@@ -9,13 +9,13 @@ summary_params = Summary_Runner()
 
 
 # TODO: Implement passing of two files to scheduled_in
-def create_summary():
+def create_summary(logger: Logger = Logger()):
     tgb.text("###### File selection (quantification)", mode="markdown")
     create_file_selection(
-        process="summary", execution_key_in="quantification", out_node="summary_data_paths"
+        process="summary", execution_key_in="quantification", out_node="summary_data_paths", logger=logger,
     )
 
     tgb.text("###### File selection (annotation)", mode="markdown")
     create_file_selection(
-        process="summary", execution_key_in="annotation", out_node="summary_data_paths"
+        process="summary", execution_key_in="annotation", out_node="summary_data_paths", logger=logger,
     )