+++ rampt/gui/pages/annotation/sirius.py
@@ -10,15 +10,15 @@ sirius_params = Sirius_Runner()
 sirius_params.config = os.path.join(ROOT_DIR, "statics", "batch_files", "sirius_default_config.txt")
 
 
-def create_sirius():
+def create_sirius(logger: Logger = Logger()):
     tgb.text("###### File selection", mode="markdown")
-    create_file_selection(process="sirius", out_node="sirius_annotation_paths")
+    create_file_selection(process="sirius", out_node="sirius_annotation_paths", logger=logger)
 
     tgb.html("br")
 
     tgb.text("###### Config selection", mode="markdown")
     create_list_selection(
-        process="sirius", attribute="config", extensions="*", name="configuration"
+        process="sirius", attribute="config", extensions="*", name="configuration",
     )
 
     create_advanced_settings()