+++ rampt/steps/analysis/summary_pipe.py
@@ -66,6 +66,7 @@ class Summary_Runner(Pipe_Step):
         save_log=False,
         additional_args: list = [],
         verbosity=1,
+        logger: Logger = Logger(),
         **kwargs,
     ):
         """
@@ -79,8 +80,10 @@ class Summary_Runner(Pipe_Step):
         :type additional_args: list, optional
         :param verbosity: Level of verbosity, defaults to 1
         :type verbosity: int, optional
+        :param logger: Logger class to handle output, defaults to Logger()
+        :type logger: Logger
         """
-        super().__init__(save_log=save_log, additional_args=additional_args, verbosity=verbosity)
+        super().__init__(save_log=save_log, additional_args=additional_args, verbosity=verbosity, logger=logger)
         if kwargs:
             self.update(kwargs)
         self.overwrite = overwrite
@@ -407,13 +410,13 @@ class Summary_Runner(Pipe_Step):
         summary.to_csv(out_path, sep="\t")
 
         if in_paths_annotation:
-            log(
+            self.logger.log(
                 f"Added annotation from {in_paths_annotation} to {in_path_quantification}.Exported to {out_path}.",
                 minimum_verbosity=1,
                 verbosity=self.verbosity,
             )
         else:
-            log(
+            self.logger.log(
                 f"No annotation received. {in_path_quantification} exported to {out_path}.",
                 minimum_verbosity=1,
                 verbosity=self.verbosity,